# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /transfers:
    post:
      tags:
      - transfer-controller
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransferResponse"
      security:
      - bearerAuth: []
  /cards/{id}/block-request:
    post:
      tags:
      - card-controller
      operationId: requestBlock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      security:
      - bearerAuth: []
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /admin/cards:
    get:
      tags:
      - admin-card-controller
      operationId: list
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - admin-card-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
      security:
      - bearerAuth: []
  /admin/cards/{id}/block-approve:
    post:
      tags:
      - admin-card-controller
      operationId: approveBlock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      security:
      - bearerAuth: []
  /admin/cards/{id}/activate:
    post:
      tags:
      - admin-card-controller
      operationId: activate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      security:
      - bearerAuth: []
  /admin/cards/{id}:
    get:
      tags:
      - admin-card-controller
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - admin-card-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      security:
      - bearerAuth: []
    patch:
      tags:
      - admin-card-controller
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
      security:
      - bearerAuth: []
  /cards:
    get:
      tags:
      - card-controller
      operationId: getUserCards
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - CREATED
          - ACTIVE
          - BLOCK_REQUESTED
          - BLOCKED
          - EXPIRED
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
      security:
      - bearerAuth: []
  /cards/{id}:
    get:
      tags:
      - card-controller
      operationId: getUserCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
      security:
      - bearerAuth: []
  /cards/{id}/balance:
    get:
      tags:
      - card-controller
      operationId: getBalance
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BalanceResponse"
      security:
      - bearerAuth: []
components:
  schemas:
    TransferRequest:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          minimum: 0.01
        differentCards:
          type: boolean
      required:
      - amount
      - fromCardId
      - toCardId
    TransferResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
    AuthRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - password
      - username
    AuthResponse:
      type: object
      properties:
        token:
          type: string
    CardCreateRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        number:
          type: string
          minLength: 1
          pattern: "\\d{16}"
        expiry:
          type: string
          format: date
      required:
      - expiry
      - number
      - userId
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        owner:
          type: string
        expiry:
          type: string
          format: date
        balance:
          type: number
        status:
          type: string
    CardUpdateRequest:
      type: object
      properties:
        expiry:
          type: string
          format: date
    BalanceResponse:
      type: object
      properties:
        balance:
          type: number
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT